- refactored PreferencesManager tests in a single class - implemented Hyperlink's actions to open their
